fix valıdatıon

+++ Services/Catalog/FreeCourse.Services.Catalog/Models/Response/FeatureResponse.cs
@@ -2,6 +2,6 @@ namespace FreeCourse.Services.Catalog.Models.Response
 {
     public class FeatureResponse
     {
-        public int Duration { get; set; }
+        public short Duration { get; set; }
     }
 }
